move abstract test classes to Test namespace
