+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -40,7 +40,7 @@
         "symfony/css-selector": "~2.8|~3.0|~4.0",
         "symfony/dom-crawler": "~2.8|~3.0|~4.0",
         "symfony/polyfill-intl-icu": "~1.0",
-        "symfony/form": "^3.4.31|^4.3.4",
+        "symfony/form": "^3.4.32|^4.3.5",
         "symfony/expression-language": "~2.8|~3.0|~4.0",
         "symfony/process": "~2.8|~3.0|~4.0",
         "symfony/security-core": "~3.2|~4.0",
