+++ src/Symfony/Component/Form/Test/AbstractLayoutTest.php
@@ -9,13 +9,11 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Form\Tests;
+namespace Symfony\Component\Form\Test;
 
 use Symfony\Component\Form\Extension\Csrf\CsrfExtension;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\Form\FormView;
-use Symfony\Component\Form\Test\FormIntegrationTestCase;
-use Symfony\Component\Form\Test\VersionAwareTest;
 
 abstract class AbstractLayoutTest extends FormIntegrationTestCase
 {
