+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/ButtonTypeTest.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
+use Symfony\Component\Form\Test\Extension\Core\Type\BaseTypeTest;
+
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
