+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/CheckboxTypeTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
 use Symfony\Component\Form\CallbackTransformer;
+use Symfony\Component\Form\Test\Extension\Core\Type\BaseTypeTest;
 
 class CheckboxTypeTest extends BaseTypeTest
 {
