+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/ChoiceTypeTest.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
 use Symfony\Component\Form\ChoiceList\View\ChoiceGroupView;
 use Symfony\Component\Form\ChoiceList\View\ChoiceView;
+use Symfony\Component\Form\Test\Extension\Core\Type\BaseTypeTest;
 
 class ChoiceTypeTest extends BaseTypeTest
 {
