+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/TextTypeTest.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
+use Symfony\Component\Form\Test\Extension\Core\Type\BaseTypeTest;
+
 class TextTypeTest extends BaseTypeTest
 {
     const TESTED_TYPE = 'Symfony\Component\Form\Extension\Core\Type\TextType';
