+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/TimeTypeTest.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
 use Symfony\Component\Form\ChoiceList\View\ChoiceView;
 use Symfony\Component\Form\FormError;
+use Symfony\Component\Form\Test\Extension\Core\Type\BaseTypeTest;
 
 class TimeTypeTest extends BaseTypeTest
 {
