+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/TimezoneTypeTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
 use Symfony\Component\Form\ChoiceList\View\ChoiceView;
+use Symfony\Component\Form\Test\Extension\Core\Type\BaseTypeTest;
 
 class TimezoneTypeTest extends BaseTypeTest
 {
