+++ src/Symfony/Component/Form/Tests/Extension/Validator/ValidatorExtensionTest.php
@@ -26,7 +26,7 @@ use Symfony\Component\Validator\Mapping\ClassMetadata;
 use Symfony\Component\Validator\Mapping\Factory\LazyLoadingMetadataFactory;
 use Symfony\Component\Validator\Mapping\Loader\StaticMethodLoader;
 use Symfony\Component\Validator\Mapping\TraversalStrategy;
-use Symfony\Component\Validator\Tests\Fixtures\FakeMetadataFactory;
+use Symfony\Component\Validator\Test\Fixtures\FakeMetadataFactory;
 use Symfony\Component\Validator\Validation;
 
 class ValidatorExtensionTest extends TestCase
