+++ src/Symfony/Component/Form/composer.json
@@ -26,7 +26,7 @@
     },
     "require-dev": {
         "doctrine/collections": "~1.0",
-        "symfony/validator": "^3.2.5|~4.0",
+        "symfony/validator": "^3.4.32|^4.3.5",
         "symfony/dependency-injection": "~3.3|~4.0",
         "symfony/config": "~2.7|~3.0|~4.0",
         "symfony/http-foundation": "~2.8|~3.0|~4.0",
